Moved metamask.ts to utils\frontend
