+++ src/components/Header.tsx
@@ -8,7 +8,7 @@ import clsx from "clsx";
 import Button from "@mui/material/Button";
 import ThemeSwitch from "../fragments/ThemeSwitch.Header";
 import { useAccount } from "../pages/_app";
-import { connected, connectWallet, network } from "../utils/metamask";
+import { connected, connectWallet, network } from "../utils/frontend/metamask";
 import { getAccountDisplayValue } from "../utils/frontend/helpers";
 
 type Route =
