+++ src/datasources/NFT/localhost.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from "ethers";
-import { getSignedContract } from "../../utils/metamask";
+import { getSignedContract } from "../../utils/frontend/metamask";
 import addresses from "../../../addresses.json";
 import { ABIs, MAX_MINTABLE_TOKENS } from "../../utils/constants";
 import { NFTAsset } from "../../models/nftAsset";
